Fixes for Java long in ObjC signatures.
git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@1260 0874c2aa-48f6-4a0f-8a6b-e32f4465a257

+++ src/xmlvm2objc/compat-lib/objc/java_io_BufferedInputStream.m
@@ -81,7 +81,7 @@
 	[target reset__];
 }
 
-- (long) skip___long: (long) n
+- (long long) skip___long: (long long) n
 {
 	return [target skip___long: n];
 }
