+++ src/xmlvm2objc/compat-lib/objc/java_io_Reader.h
@@ -39,7 +39,7 @@
 - (bool) ready__;
 - (void) mark___int: (int) readAheadLimit;
 - (BOOL) markSupported__;
-- (long) skip___long: (long) n;
+- (long long) skip___long: (long) n;
 - (void) reset__;
 - (void) close__;
