+++ src/xmlvm2objc/compat-lib/objc/java_lang_System.m
@@ -49,7 +49,7 @@ java_io_InputStream* _STATIC_java_lang_System_in;
  value it doesn't return an accurate time according to the Java spec for this method.
  Problems lie in the long representation on the 32 bit target architecture.
  */
-+ (long) currentTimeMillis__
++ (long long) currentTimeMillis__
 {
     static NSDate* now_time;
     static long    msec_since_1970;
@@ -63,7 +63,7 @@ java_io_InputStream* _STATIC_java_lang_System_in;
     return sinceNow + msec_since_1970;
 }
 
-+ (long) nanoTime__
++ (long long) nanoTime__
 {
     static NSDate* now_time;
     
@@ -74,7 +74,7 @@ java_io_InputStream* _STATIC_java_lang_System_in;
     time *= -1000.0; // milli
     time *= 1000.0; // micro
     time *= 1000.0; // nano
-    return (long) floor(time + 0.5);
+    return (long long) floor(time + 0.5);
 }
 
 + (void) initialize
